Added Logout button
* Button will be visible for laptop view
* Icon will be there for mobile view
