+++ src/components/SideBar/NavBar.jsx
@@ -2,22 +2,10 @@ import { signOut } from "firebase/auth";
 import React, { useContext } from "react";
 import { AuthContext } from "../../context/AuthContext";
 import { auth, db } from "../../firebase";
-import moreIcon from "../../img/logout.png";
 import PlaceholderLoading from "react-placeholder-loading";
-import { doc, Timestamp, updateDoc } from "firebase/firestore";
 
 const NavBar = () => {
-  var isOfflineForDatabase = {
-    state: "offline",
-    last_changed: Timestamp.now().seconds,
-  };
   const { currentUser } = useContext(AuthContext);
-  const handleSignOut = async () => {
-    await updateDoc(doc(db, "/users/", currentUser.uid), {
-      status: isOfflineForDatabase
-    });
-    signOut(auth);
-  };
 
   return (
     <div className="navBar">
@@ -42,18 +30,6 @@ const NavBar = () => {
           )}
         </div>
       </span>
-      <img
-        src={moreIcon}
-        className="moreIcon"
-        style={{
-          width: "25px",
-          height: "25px",
-          backgroundColor: "transparent",
-          cursor: "pointer",
-        }}
-        alt=""
-        onClick={handleSignOut}
-      />
     </div>
   );
 };
